==> 1_data_process.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn import random_projection
import lightgbm as lgb
from math import exp, log
path = '/home/ec2-user/Sanjay/main/santander/'

train_df = pd.read_csv(path + 'train.csv')
test_df = pd.read_csv(path + 'test.csv')

X_train = train_df.drop(["ID", "target"], axis=1)
y_train = train_df[['target']]['target'].map(lambda x : log(1+x))

X_test = test_df.drop(["ID"], axis=1)
ntrain = len(X_train)
ntest = len(X_test)
X_train.shape, X_test.shape

print("Preparetion")
colsToRemove = []
for col in X_train.columns:
    if X_train[col].std() == 0: 
        colsToRemove.append(col)
    
X_train.drop(colsToRemove, axis=1, inplace=True)
X_test.drop(colsToRemove, axis=1, inplace=True)

X_train.shape, X_test.shape, len(colsToRemove)

colsToRemove = []
colsScaned = []
dupList = {}
columns = X_train.columns

for i in range(len(columns)-1):
    v = X_train[columns[i]].values
    dupCols = []
    for j in range(i+1,len(columns)):
        if np.array_equal(v, X_train[columns[j]].values):
            colsToRemove.append(columns[j])
            if columns[j] not in colsScaned:
                dupCols.append(columns[j]) 
                colsScaned.append(columns[j])
                dupList[columns[i]] = dupCols
                
X_train.drop(colsToRemove, axis=1, inplace=True) 
X_test.drop(colsToRemove, axis=1, inplace=True)
X_train.shape, X_test.shape, len(dupList)

X_train.to_csv(path + 'codes/data/' + '1_raw_data_train.csv', index=False)
X_test.to_csv(path + 'codes/data/' + '1_raw_data_test.csv', index=False)
