==> 2_Random_Projection.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn import random_projection
import lightgbm as lgb
from math import exp, log
path = '/home/ec2-user/Sanjay/main/santander/'

X_train = pd.read_csv(path + 'codes/data/' + '1_raw_data_train.csv')
X_test = pd.read_csv(path + 'codes/data/' + '1_raw_data_test.csv')
ntrain = len(X_train)
ntest = len(X_test)
X_train.shape, X_test.shape

tmp = pd.concat([X_train,X_test]).fillna(0)
tmp.shape

n_com = 50
transformer = random_projection.SparseRandomProjection(n_components = n_com)
RP = transformer.fit_transform(tmp)
rp = pd.DataFrame(RP)
columns = ["RandomProjection{}".format(i) for i in range(n_com)]
rp.columns = columns

rp_train = rp[:ntrain]
rp_test = rp[ntrain:]
rp_test.index = X_test.index

rp_train.shape, rp_test.shape

rp_train.to_csv(path + 'codes/data/' + '2_Random_Projection_train.csv', index=False)
rp_test.to_csv(path + 'codes/data/' + '2_Random_Projection_test.csv', index=False)
