==> 4_aggregated.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn import random_projection
import lightgbm as lgb
from math import exp, log
path = '/home/ec2-user/Sanjay/main/santander/'

X_train = pd.read_csv(path + 'codes/data/' + '1_raw_data_train.csv')
X_test = pd.read_csv(path + 'codes/data/' + '1_raw_data_test.csv')
X_train.shape, X_test.shape

weight = ((X_train != 0).sum()/len(X_train)).values

tmp_train = X_train[X_train!=0]
tmp_test = X_test[X_test!=0]
tmp = pd.concat([X_train,X_test]).fillna(0)
tmp_train.shape, tmp_test.shape, tmp.shape

X_train["weight_count"] = (tmp_train*weight).sum(axis=1)
X_test["weight_count"] = (tmp_test*weight).sum(axis=1)
X_train["count_0"] = (X_train == 0).sum(axis=1)
X_test["count_0"] = (X_test == 0).sum(axis=1)
X_train["count_not0"] = (X_train != 0).sum(axis=1)
X_test["count_not0"] = (X_test != 0).sum(axis=1)
X_train["sum"] = X_train.sum(axis=1)
X_test["sum"] = X_test.sum(axis=1)
X_train["var"] = tmp_train.var(axis=1)
X_test["var"] = tmp_test.var(axis=1)
X_train["mean"] = tmp_train.mean(axis=1)
X_test["mean"] = tmp_test.mean(axis=1)
X_train["std"] = tmp_train.std(axis=1)
X_test["std"] = tmp_test.std(axis=1)
X_train["max"] = tmp_train.max(axis=1)
X_test["max"] = tmp_test.max(axis=1)
X_train["min"] = tmp_train.min(axis=1)
X_test["min"] = tmp_test.min(axis=1)

X_train.iloc[:,4730:].to_csv(path + 'codes/data/' + '4_aggr_train.csv', index=False)
X_test.iloc[:,4730:].to_csv(path + 'codes/data/' + '4_aggr_test.csv', index=False)
