==> 5_pca.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn import random_projection
import lightgbm as lgb
from math import exp, log
from sklearn import model_selection
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import RobustScaler
from sklearn.decomposition import TruncatedSVD
from sklearn.preprocessing import StandardScaler
from sklearn.manifold import TSNE
from sklearn.cluster import KMeans
from sklearn.decomposition import PCA
from sklearn.preprocessing import normalize
path = '/home/ec2-user/Sanjay/main/santander/'

X_train = pd.read_csv(path + 'codes/data/' + '1_raw_data_train.csv')
X_test = pd.read_csv(path + 'codes/data/' + '1_raw_data_test.csv')
X_train = X_train.fillna(0)
X_test = X_test.fillna(0)
X_train.shape, X_test.shape

flist = [x for x in X_train.columns]

n_components = 20
flist_pca = []
pca = PCA(n_components=n_components)
x_train_projected = pca.fit_transform(normalize(X_train[flist], axis=0))
x_test_projected = pca.transform(normalize(X_test[flist], axis=0))
for npca in range(0, n_components):
    X_train.insert(1, 'PCA_'+str(npca+1), x_train_projected[:, npca])
    X_test.insert(1, 'PCA_'+str(npca+1), x_test_projected[:, npca])
    flist_pca.append('PCA_'+str(npca+1))
    print(flist_pca)

X_train[flist_pca].to_csv(path + 'codes/data/' + '5_pca_train.csv', index=False)
X_test[flist_pca].to_csv(path + 'codes/data/' + '5_pca_test.csv', index=False)
