==> 6_Model_lgbm.ipynb <==
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn import random_projection
import lightgbm as lgb
from math import exp, log
path = '/home/ec2-user/Sanjay/main/santander/'

X_train = pd.read_csv(path + 'codes/data/' + '1_raw_data_train.csv')
X_test = pd.read_csv(path + 'codes/data/' + '1_raw_data_test.csv')

test_df = pd.read_csv(path + 'test.csv',  usecols=['ID'])
train_df = pd.read_csv(path + 'train.csv', usecols=['target'])
y_train = train_df[['target']]['target'].map(lambda x : log(1+x))
X_train.shape, X_test.shape, train_df.shape

Random_Projection_train = pd.read_csv(path + 'codes/data/' + '2_Random_Projection_train.csv')
Random_Projection_test = pd.read_csv(path + 'codes/data/' + '2_Random_Projection_test.csv')

kmeans_train = pd.read_csv(path + 'codes/data/' + '3_kmeans_train.csv')
kmeans_test = pd.read_csv(path + 'codes/data/' + '3_kmeans_test.csv')

aggr_train = pd.read_csv(path + 'codes/data/' + '4_aggr_train.csv')
aggr_test = pd.read_csv(path + 'codes/data/' + '4_aggr_test.csv')

pca_train = pd.read_csv(path + 'codes/data/' + '5_pca_train.csv')
pca_test = pd.read_csv(path + 'codes/data/' + '5_pca_test.csv')

print(Random_Projection_train.shape, Random_Projection_test.shape, 
      kmeans_train.shape, kmeans_test.shape,
      aggr_train.shape, aggr_test.shape, 
      pca_train.shape, pca_test.shape )

train = X_train.join(Random_Projection_train).join(kmeans_train).join(aggr_train).join(pca_train)
test = X_test.join(Random_Projection_test).join(kmeans_test).join(aggr_test).join(pca_test)
train.shape, test.shape

#### 10 fold CV LightGBM

parameters = {
    'task': 'train',
    'boosting_type': 'gbdt',
    'objective': 'regression',
    'metric': 'rmse',
    'num_leaves': 100,
    'learning_rate': 0.005,
    'feature_fraction': 0.9,
    'bagging_fraction': 0.8,
    'bagging_seed' : 42,
    'bagging_freq': 5,
    'verbose': 50
}

def lgbm_model(X_tr, y_tr, X_va, y_va, test, fold):
    tr_data = lgb.Dataset(X_tr, label=y_tr)
    va_data = lgb.Dataset(X_va, label=y_va)
    
    model = lgb.train(parameters,
                      tr_data,
                      valid_sets=va_data,
                      num_boost_round=4000,
                      early_stopping_rounds=120,
                      verbose_eval=300)
    
    y_pred = model.predict(X_va)
    valid = pd.DataFrame(y_pred, columns=['prob' + str(fold)])
    test_pred = model.predict(test)
    test = pd.DataFrame(test_pred, columns=['prob' + str(fold)])
    return test

from sklearn.model_selection import KFold
X = train
y = y_train
kf = KFold(n_splits=10, shuffle=True, random_state=2017)
kf.get_n_splits(X), kf.split(X)
fold=1
valid_final = test_df[['ID']]
for train_idx, test_idx in kf.split(X):
    print("FOLD ", fold, " ----------------------------------------------------------------------------------")
    X_tr = X[X.index.isin(train_idx)]
    y_tr = y[y.index.isin(train_idx)]
    X_va = X[X.index.isin(test_idx)]
    y_va = y[y.index.isin(test_idx)]
    print(X_tr.shape, y_tr.shape, X_va.shape, y_va.shape)
    valid = lgbm_model(X_tr, y_tr, X_va, y_va, test, fold)
    valid_final = valid_final.join(valid)
    print(valid.shape, valid_final.shape)
    fold = fold+1

valid_final['target'] = valid_final.iloc[:,1:fold].mean(axis=1)
valid_final['target'] = valid_final['target'].map(lambda x : exp(x)-1)

valid_final = valid_final.iloc[:49342]
valid_final.to_csv(path + 'submit/' + 'lgbm_10fold_1.csv', index=False)
valid_final[['ID', 'target']].to_csv(path + 'submit/' + 'lgbm_1.csv', index=False)

### Predict the nearby input

pred = pd.read_csv(path + 'submit/' + 'lgbm_1.csv')
pred_input = pd.read_csv(path + 'test.csv')
pred = pred.merge(pred_input, on='ID', how='inner')
pred.shape, pred_input.shape

def nearby_value(array, value):
    idx,val = min(enumerate(np.sort(array[array!=0])), key=lambda x: abs(x[1]-value))
    return val

close_value = []
for i in range(pred.shape[0]):
    if (i%5000)==0:
        print(i)
    close_value.append(nearby_value(pred.iloc[0][pred.columns[1:]].values, pred.target[i]))

pred['target'] = close_value

pred[['ID', 'target']].to_csv(path + 'submit/' + 'lgbm_1_nearby_input.csv', index=False)

### Ensemble

import pandas as pd
path = '/home/ec2-user/Sanjay/main/santander/'
data1 = pd.read_csv(path + 'submit/' + 'baseline_sub_lag_37.csv')
data2 = pd.read_csv(path + 'submit/' + 'check_blend1.csv')

import copy
data = data1.copy()
data['target'] = data1['target']*0.7 + data2['target']*0.3

data[['ID', 'target']].to_csv(path + 'submit/' + 'check_blend1.csv', index=False)
